+++ network/tests/fixtures/mac_netstat
@@ -1,0 +1,42 @@
+Name       Mtu   Network       Address            Ipkts Ierrs     Ibytes    Opkts Oerrs     Obytes  Coll
+lo0        16384 <Link#1>                         91928     0   17326550    91928     0   17326550     0
+lo0        16384 127           localhost          91928     -   17326550    91928     -   17326550     -
+lo0        16384 localhost   ::1                  91928     -   17326550    91928     -   17326550     -
+lo0        16384 comp-wv79yp fe80:1::1            91928     -   17326550    91928     -   17326550     -
+gif0*      1280  <Link#2>                             0     0          0        0     0          0     0
+stf0*      1280  <Link#3>                             0     0          0        0     0          0     0
+anpi2      1500  <Link#4>    ce:69:62:2f:26:00        0     0          0        0     0          0     0
+anpi2      1500  comp-wv79yp fe80:4::cc69:62ff        0     -          0        0     -          0     -
+anpi1      1500  <Link#5>    ce:69:62:2f:26:ff        0     0          0        0     0          0     0
+anpi1      1500  comp-wv79yp fe80:5::cc69:62ff        0     -          0        0     -          0     -
+anpi0      1500  <Link#6>    ce:69:62:2f:26:fe        0     0          0        0     0          0     0
+anpi0      1500  comp-wv79yp fe80:6::cc69:62ff        0     -          0        0     -          0     -
+en4        1500  <Link#7>    ce:69:62:2f:26:de        0     0          0        0     0          0     0
+en5        1500  <Link#8>    ce:69:62:2f:26:df        0     0          0        0     0          0     0
+en6        1500  <Link#9>    ce:69:62:2f:26:e0        0     0          0        0     0          0     0
+en1        1500  <Link#10>   36:e4:12:d3:f9:40        0     0          0        0     0          0     0
+en2        1500  <Link#11>   36:e4:12:d3:f9:44        0     0          0        0     0          0     0
+en3        1500  <Link#12>   36:e4:12:d3:f9:48        0     0          0        0     0          0     0
+ap1        1500  <Link#13>   be:d0:74:21:5b:76        0     0          0        0     0          0     0
+en0        1500  <Link#14>   bc:d0:74:21:5b:76  5473980     0 4404763710  3683504     0 2235748839     0
+en0        1500  comp-wv79yp fe80:e::1c39:bbc7  5473980     - 4404763710  3683504     - 2235748839     -
+en0        1500  192.168.135   192.168.135.119  5473980     - 4404763710  3683504     - 2235748839     -
+awdl0      1500  <Link#15>   02:a8:81:d8:41:ad      197     0      73861      597     0     111052     0
+awdl0      1500  fe80::a8:81 fe80:f::a8:81ff:f      197     -      73861      597     -     111052     -
+llw0       1500  <Link#16>   02:a8:81:d8:41:ad        0     0          0        0     0          0     0
+llw0       1500  fe80::a8:81 fe80:10::a8:81ff:        0     -          0        0     -          0     -
+bridge0    1500  <Link#17>   36:e4:12:d3:f9:40        0     0          0        0     0          0     0
+utun0      1380  <Link#18>                            0     0          0      726     0      39296     0
+utun0      1380  comp-wv79yp fe80:12::df2:23aa        0     -          0      726     -      39296     -
+utun1      2000  <Link#19>                            0     0          0      726     0      39296     0
+utun1      2000  comp-wv79yp fe80:13::e784:e85        0     -          0      726     -      39296     -
+utun2      1000  <Link#20>                            0     0          0      726     0      39296     0
+utun2      1000  comp-wv79yp fe80:14::ce81:b1c        0     -          0      726     -      39296     -
+utun3      1359  <Link#21>                            0     0          0        2     0        240     0
+utun3      1359  192.168.130.2 192.168.130.254        0     -          0        2     -        240     -
+utun3      1359  comp-wv79yp fe80:15::bed0:74f        0     -          0        2     -        240     -
+utun3      1359  fd15:c974:f fd15:c974:fd30:aa        0     -          0        2     -        240     -
+utun4      1380  <Link#22>                           26     0       3876       31     0       3490     0
+utun4      1380  comp-wv79yp fe80:16::87e6:73b       26     -       3876       31     -       3490     -
+utun5      1380  <Link#23>                            0     0          0       63     0       3494     0
+utun5      1380  comp-wv79yp fe80:17::c1c5:17a        0     -          0       63     -       3494     -
