+++ network/tests/test_integration.py
@@ -5,7 +5,6 @@ import platform
 
 import pytest
 
-from datadog_checks.network.check_bsd import BSDNetwork
 from . import common
 
 pytestmark = pytest.mark.integration
@@ -27,13 +26,4 @@ def test_check_linux(aggregator, check, instance_blacklist):
     check_instance.check({})
 
     for metric in common.CONNTRACK_METRICS:
-        aggregator.assert_metric(metric)
-
-
-@pytest.mark.skipif(platform.system() == 'Windows', reason="Only runs on Unix systems")
-@pytest.mark.usefixtures("dd_environment")
-def test_check_bsd(instance, aggregator):
-    check = BSDNetwork('network', {}, [instance])
-    check.check({})
-    for metric in common.EXPECTED_METRICS:
         aggregator.assert_metric(metric)
